refactor app logic into bothan file
